refactor: generic function to validate moderation
